move allocate tranche payout to utils
