+++ climada_petals/engine/cat_bonds/mlt_bond_simulation.py
@@ -2,7 +2,7 @@ import pandas as pd
 import numpy as np
 import logging
 
-from utils_cat_bonds import multi_level_es
+from utils_cat_bonds import multi_level_es, allocate_single_payout
 
 LOGGER = logging.getLogger(__name__)
 
@@ -476,34 +476,3 @@ class MultiCountryBondSimulation:
         tot_loss = np.sum(ann_loss)
         return tot_loss
     
-def allocate_single_payout(payout, nominals):
-    """
-    Vectorised allocation of one payout across tranche nominals (FIFO).
-    
-    Parameters
-    ----------
-    payout : float
-    nominals : 1D array of tranche nominal values
-
-    Returns
-    -------
-    alloc : array of size (T,)  -- how much each tranche pays
-    remaining_nominals : array -- leftover nominals after the payout
-    """
-
-    nominals = np.asarray(nominals, float)
-
-    # cumulative nominal capacity per tranche
-    cum_nom = np.cumsum(nominals)
-    cum_nom_prev = cum_nom - nominals
-
-    # intersection of [0, payout] with each tranche interval [cum_nom_prev, cum_nom]
-    payout_per_tranche = np.minimum(cum_nom, payout) - np.maximum(cum_nom_prev, 0)
-
-    # clip negative / unused intervals
-    payout_per_tranche = np.clip(payout_per_tranche, 0, None)
-
-    remaining_nominals = nominals - payout_per_tranche
-
-
-    return remaining_nominals, payout_per_tranche
