+++ climada_petals/engine/cat_bonds/utils_cat_bonds.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 '''Calculate value at risk and expected shorfall for various alphas'''
 def multi_level_es(losses, confidence_levels):
     """
@@ -21,3 +23,36 @@ def multi_level_es(losses, confidence_levels):
     ]
 
     return var_list, es_list
+
+
+def allocate_single_payout(payout, nominals):
+    """
+    Vectorised allocation of one payout across tranche nominals (FIFO).
+    
+    Parameters
+    ----------
+    payout : float
+    nominals : 1D array of tranche nominal values
+
+    Returns
+    -------
+    alloc : array of size (T,)  -- how much each tranche pays
+    remaining_nominals : array -- leftover nominals after the payout
+    """
+
+    nominals = np.asarray(nominals, float)
+
+    # cumulative nominal capacity per tranche
+    cum_nom = np.cumsum(nominals)
+    cum_nom_prev = cum_nom - nominals
+
+    # intersection of [0, payout] with each tranche interval [cum_nom_prev, cum_nom]
+    payout_per_tranche = np.minimum(cum_nom, payout) - np.maximum(cum_nom_prev, 0)
+
+    # clip negative / unused intervals
+    payout_per_tranche = np.clip(payout_per_tranche, 0, None)
+
+    remaining_nominals = nominals - payout_per_tranche
+
+
+    return remaining_nominals, payout_per_tranche
